feat(red_tag): add red tag options for coupling beam handling
Introduce options to control the tagging behavior of coupling beams in the red tag assessment. This includes the ability to ignore coupling beams
for red tag calculations and to tag them based on height. These options
enhance the flexibility of the red tag functionality.

Update the `fn_red_tag` function to accept `red_tag_options` and modify
the tagging logic accordingly. Also, update the `optional_inputs` to
include these new options.

This feature update is in line with commit F6ea7c6 of the matlab recover code. The red tag logic for coupling beams is a bit of a patch and should be revisited later for further refactoring.